get_all_minidata.py - allow selecting by DPRTYPE + DRSOUTID + fiber

+++ general/apero_get/get_all_minidata.py
@@ -22,14 +22,32 @@ from tqdm import tqdm
 workdir = '/scratch2/drs-data/'
 # define where to copy these files to
 output = '/spirou2/minidata/'
+# reset
+RESET = False
 # define the targets (as in DRSOBJN)
 # targets = ['GL699']
 # targets = None --> all targets
 targets = None
-# define file types (glob path with wildcards from the reduced directory)
-filetypes = ['*/2*o_pp_e2dsff_AB.fits', '*/2*o_pp_e2dsff_C.fits',
-             '*/2*o_pp_e2dsff_tcorr_AB.fits',
-             'other/Template_s1d_*AB.fits', 'other/Template_s1d_*C.fits']
+
+# define what type of files to get
+outtypes = ['EXT_E2DS_FF', 'TELLU_OBJ', 'TELLU_TEMP', 'TELLU_TEMP_S1D']
+fibers = ['AB', 'C']
+dprtypes = ['OBJ_FP', 'OBJ_DARK', 'POLAR_FP', 'POLAR_DARK', 'FP_FP']
+
+
+def get_fiber(header, filename):
+    if 'FIBER' in header:
+        return str(header['FIBER'])
+    if 'AB.fits' in filename:
+        return 'AB'
+    elif 'A.fits' in filename:
+        return 'A'
+    elif 'B.fits' in filename:
+        return 'B'
+    elif 'C.fits' in filename:
+        return 'C'
+    else:
+        return 'NULL'
 
 
 # =============================================================================
@@ -45,16 +63,14 @@ if __name__ == "__main__":
         # get outdir
         outdir = os.path.join(output, os.path.basename(directory))
         # rm outdir if it exists
-        if os.path.exists(outdir):
+        if os.path.exists(outdir) and RESET:
             os.system('rm -rfv {0}'.format(outdir))
-        # make out dir
-        os.mkdir(outdir)
+            # make out dir
+            os.mkdir(outdir)
         # get pattern
         pattern = directory + '/reduced/'
         # get files
-        all_files = []
-        for filetype in filetypes:
-            all_files += glob.glob(pattern + filetype)
+        all_files = glob.glob(pattern + '*/*.fits')
         # loop around files and add to outdir (if correct target)
         for filename in tqdm(all_files):
             # get header for this file
@@ -62,12 +78,30 @@ if __name__ == "__main__":
             if 'DRSOBJN' in header:
                 objname = str(header['DRSOBJN'])
             else:
-                header = fits.getheader(filename, ext=1)
-                objname = str(header['DRSOBJN'])
+                try:
+                    header = fits.getheader(filename, ext=1)
+                    objname = str(header['DRSOBJN'])
+                # skip if we still don't have object name
+                except Exception as _:
+                    continue
+            # get dprtype fiber and outtype
+            dprtype = str(header.get('DPRTYPE', 'NULL'))
+            fiber = get_fiber(header, filename)
+            outtype = str(header.get('DRSOUTID', 'NULL'))
+            # remove header
             del header
             # do not continue if objname not in targets
             if targets is not None:
                 if objname not in targets:
+                    continue
+            if dprtypes is not None:
+                if dprtype not in dprtypes:
+                    continue
+            if fibers is not None:
+                if fiber not in fibers:
+                    continue
+            if outtypes is not None:
+                if outtype not in outtypes:
                     continue
             # reate out path
             outpath = os.path.join(outdir, objname)
@@ -77,7 +111,8 @@ if __name__ == "__main__":
             # construct outpath
             outfilename = os.path.join(outpath, os.path.basename(filename))
             # copy
-            shutil.copy(filename, outfilename)
+            if not os.path.exists(outfilename):
+                shutil.copy(filename, outfilename)
 
 # =============================================================================
 # End of code
